requirements.txt
Flask
requests
python-dotenv
requests-retry
gunicorn
urllib3
